cleaned up output

+++ commands/src/main/kotlin/com/seanshubin/project/generator/commands/WriteTextFile.kt
@@ -1,6 +1,8 @@
 package com.seanshubin.project.generator.commands
 
 import java.nio.file.Path
+import java.nio.file.attribute.PosixFilePermission
+import java.nio.file.attribute.PosixFilePermission.*
 
 data class WriteTextFile(val path: Path, val content: String, val executable: Boolean = false) : Command {
     override fun execute(environment: Environment) {
@@ -8,17 +10,33 @@ data class WriteTextFile(val path: Path, val content: String, val executable: Bo
         if (parent != null) {
             environment.files.createDirectories(parent)
         }
-        environment.files.writeString(path, content)
-        if (executable) {
-            environment.files.setPosixFilePermissions(path, setOf(
-                java.nio.file.attribute.PosixFilePermission.OWNER_READ,
-                java.nio.file.attribute.PosixFilePermission.OWNER_WRITE,
-                java.nio.file.attribute.PosixFilePermission.OWNER_EXECUTE,
-                java.nio.file.attribute.PosixFilePermission.GROUP_READ,
-                java.nio.file.attribute.PosixFilePermission.GROUP_EXECUTE,
-                java.nio.file.attribute.PosixFilePermission.OTHERS_READ,
-                java.nio.file.attribute.PosixFilePermission.OTHERS_EXECUTE
-            ))
+
+        val existed = environment.files.exists(path)
+        if (existed) {
+            val existingContent = environment.files.readString(path)
+            if (existingContent == content) {
+                environment.fileOperationNotifications.fileUnchanged(path)
+                return
+            }
+            environment.files.writeString(path, content)
+            if (executable) {
+                environment.files.setPosixFilePermissions(path, executablePermissions)
+            }
+            environment.fileOperationNotifications.fileModified(path)
+        } else {
+            environment.files.writeString(path, content)
+            if (executable) {
+                environment.files.setPosixFilePermissions(path, executablePermissions)
+            }
+            environment.fileOperationNotifications.fileCreated(path)
         }
+    }
+
+    companion object {
+        private val executablePermissions = setOf(
+            OWNER_READ, OWNER_WRITE, OWNER_EXECUTE,
+            GROUP_READ, GROUP_EXECUTE,
+            OTHERS_READ, OTHERS_EXECUTE
+        )
     }
 }
